notification-service/requirements.txt
kafka-python 
faker 
